Created a Post class. This class will ultimately represent a post record from our database. The class has private properties and getters and setters for a title and body.
Created views for viewing all the posts and for viewing one individual post.

Edited  PostsController class to return the views created above.

Inside the method that shows an individual post, created a new post object and pass it to the view.
Inside the method that shows all the posts, created a new array list and add two post objects to it, then pass that list to the view.